Split page vote to component
