+++ framerail/src/routes/[slug]/[...extra]/index.ts
@@ -2,3 +2,4 @@ export { default as EditorPane } from "./EditorPane.svelte"
 export { default as LayoutPane } from "./LayoutPane.svelte"
 export { default as MovePane } from "./MovePane.svelte"
 export { default as ParentPane } from "./ParentPane.svelte"
+export { default as VotePane } from "./VotePane.svelte"
