+++ framerail/src/routes/[slug]/[...extra]/page.svelte
@@ -4,7 +4,7 @@
   import { onMount } from "svelte"
   import { useErrorPopup, usePagePaneState } from "$lib/stores"
   import { PagePane } from "$lib/types"
-  import { EditorPane, LayoutPane, MovePane, ParentPane } from "."
+  import { EditorPane, LayoutPane, MovePane, ParentPane, VotePane } from "."
   let showErrorPopup = useErrorPopup()
   let pagePaneState = usePagePaneState()
 
@@ -12,8 +12,6 @@
   let showSource = false
   let showRevision = false
   let showRevisionSource = false
-  let showVote = false
-  let showVoteList = false
   let showFiles = false
   let showFileUploadAction = false
   let showFileEditAction = false
@@ -22,8 +20,6 @@
   let showFileHistory = false
   let revisionMap: Map<number, Record<string, any>> = new Map()
   let revision: Record<string, any> = {}
-  let voteMap: Map<number, Record<string, any>> = new Map()
-  let voteRating: number
   let fileMap: Map<number, Record<string, any>> = new Map()
   let filesUpload: FileList
   let filesEditElem: HTMLInputElement
@@ -141,83 +137,6 @@
         data: res.data
       })
     } else invalidateAll()
-  }
-
-  async function handleVote() {
-    let fdata = new FormData()
-    fdata.set("site-id", $page.data.site.site_id)
-    fdata.set("page-id", $page.data.page.page_id)
-    let res = await fetch(`/${$page.data.page.slug}/score`, {
-      method: "POST",
-      body: fdata
-    }).then((res) => res.json())
-    if (res?.message) {
-      showErrorPopup.set({
-        state: true,
-        message: res.message,
-        data: res.data
-      })
-    } else {
-      voteRating = res.score ?? 0
-      showVote = true
-    }
-  }
-  async function getVoteList() {
-    let fdata = new FormData()
-    fdata.set("site-id", $page.data.site.site_id)
-    fdata.set("page-id", $page.data.page.page_id)
-    fdata.set("action", "get_list")
-    let res = await fetch(`/${$page.data.page.slug}/vote`, {
-      method: "POST",
-      body: fdata
-    }).then((res) => res.json())
-    if (res?.message) {
-      showErrorPopup.set({
-        state: true,
-        message: res.message,
-        data: res.data
-      })
-    } else {
-      voteMap = new Map()
-      res.forEach((vote) => {
-        voteMap.set(vote.user_id, vote)
-      })
-    }
-  }
-  async function castVote(value?: number) {
-    let fdata = new FormData()
-    fdata.set("site-id", $page.data.site.site_id)
-    fdata.set("page-id", $page.data.page.page_id)
-    fdata.set("action", "set")
-    fdata.set("value", value ?? 0)
-    let res = await fetch(`/${$page.data.page.slug}/vote`, {
-      method: "POST",
-      body: fdata
-    }).then((res) => res.json())
-    if (res?.message) {
-      showErrorPopup.set({
-        state: true,
-        message: res.message,
-        data: res.data
-      })
-    }
-  }
-  async function cancelVote() {
-    let fdata = new FormData()
-    fdata.set("site-id", $page.data.site.site_id)
-    fdata.set("page-id", $page.data.page.page_id)
-    fdata.set("action", "remove")
-    let res = await fetch(`/${$page.data.page.slug}/vote`, {
-      method: "POST",
-      body: fdata
-    }).then((res) => res.json())
-    if (res?.message) {
-      showErrorPopup.set({
-        state: true,
-        message: res.message,
-        data: res.data
-      })
-    }
   }
 
   async function getFileList(deleted?: boolean) {
@@ -534,7 +453,9 @@
     <button
       class="action-button button-vote clickable"
       type="button"
-      on:click={handleVote}
+      on:click={() => {
+        pagePaneState.set(PagePane.Vote)
+      }}
     >
       {$page.data.internationalization?.vote}
     </button>
@@ -562,6 +483,8 @@
   <LayoutPane />
 {:else if $pagePaneState === PagePane.Parent}
   <ParentPane />
+{:else if $pagePaneState === PagePane.Vote}
+  <VotePane />
 {/if}
 
 {#if showHistory}
@@ -648,72 +571,6 @@
   {#if showRevisionSource}
     <textarea class="revision-source" readonly={true}>{revision.wikitext}</textarea>
   {/if}
-{/if}
-
-{#if showVote}
-  <div class="vote-panel">
-    <div class="action-row vote-action">
-      <button
-        class="action-button view-vote-list clickable"
-        type="button"
-        on:click|stopPropagation={() => {
-          getVoteList().then(() => {
-            showVoteList = true
-          })
-        }}
-      >
-        {$page.data.internationalization?.["wiki-page-vote-list"]}
-      </button>
-      <div class="action-button vote-rating">
-        <span class="vote-desc"
-          >{$page.data.internationalization?.["wiki-page-vote-score"]}</span
-        >
-        <span class="vote-rating-number">{voteRating}</span>
-      </div>
-      <div class="action-button cast-vote">
-        <span class="vote-desc"
-          >{$page.data.internationalization?.["wiki-page-vote-set"]}</span
-        >
-        <button
-          class="vote-subbutton clickable"
-          type="button"
-          on:click|stopPropagation={() => castVote(1)}
-        >
-          +1
-        </button>
-        <button
-          class="vote-subbutton clickable"
-          type="button"
-          on:click|stopPropagation={() => castVote(0)}
-        >
-          0
-        </button>
-        <button
-          class="vote-subbutton clickable"
-          type="button"
-          on:click|stopPropagation={() => castVote(-1)}
-        >
-          -1
-        </button>
-      </div>
-      <button
-        class="action-button remove-vote clickable"
-        type="button"
-        on:click|stopPropagation={cancelVote}
-      >
-        {$page.data.internationalization?.["wiki-page-vote-remove"]}
-      </button>
-    </div>
-    {#if showVoteList}
-      <ul class="vote-list">
-        {#each [...voteMap].sort((a, b) => b[0] - a[0]) as [_, vote] (vote.page_vote_id)}
-          <li class="vote-item" data-id={vote.page_vote_id} data-user-id={vote.user_id}>
-            UT: User {vote.user_id}: {vote.value}
-          </li>
-        {/each}
-      </ul>
-    {/if}
-  </div>
 {/if}
 
 {#if showFiles}
